==> Notebook/2.character_preprocessing.ipynb <==
from functions import *

df_char = pd.read_csv('../Data/raw_data/Characters.csv', sep=';', encoding='unicode_escape')

We build a new class for those who aren't in Hogwart's houses.

houses = ['Gryffindor', 'Ravenclaw', 'Slytherin', 'Hufflepuff']
for i in range(len(df_char['House'])):
    if df_char['House'][i] not in houses:
        df_char['House'][i] = 'Other'

We delete features we don't want to work with (which seem to be useless).

df_char = df_char.drop(["Job", "Wand", "Patronus", "Species", "Hair colour", "Eye colour", "Birth", "Death"], axis=1)

# df_char.to_csv('../Data/temporary_data/Characters.csv', sep=';', index=False)

After exporting it, we modify the raw file to make charcaters' name fit those in the other dataframes.  
  
Example :  
 * Harry Potter beacames Harry
 * Ginevra (Ginny) Molly Weasley beacames Ginny

Then we reorder skills and loyalty columns as lists.


Finally we make a list containig the new characters appearing in the third film and a new table with them.

df_1_2 = pd.read_csv('../Data/temporary_data/Words_by_character_modified.csv', sep=';', encoding='unicode_escape')

df3 = pd.read_csv('../Data/temporary_data/Words_by_character_3_modified.csv', sep=';', encoding='unicode_escape')

list_character = df_1_2.Character.values.tolist()

list_character_3 = df3.Character.values.tolist()

list_character = [name.lower() for name in list_character]
list_character = [name.strip() for name in list_character]

list_character_3 = [name.lower() for name in list_character_3]
list_character_3 = [name.strip() for name in list_character_3]

new_character_3 = []
for character in list_character_3:
    if character not in list_character:
        new_character_3.append(character)
new_character_3 = [string.capwords(name) for name in new_character_3]

Here is the list of new characters talking in the third film:

new_character_3

df_characters = pd.read_csv('../Data/temporary_data/Characters_modified.csv', sep=';', encoding='unicode_escape')

df_char_3 = df_characters.copy()
# we make a new df with only the new characters appearing in hp3 whom informations are given
for i in range(len(df_char_3)):
    if df_char_3.Character[i] not in new_character_3:
        df_char_3 = df_char_3.drop([i])

df_char_3 = df_char_3.reset_index(drop=True)

# df_char_3.to_csv('../Data/temporary_data/New_characters.csv', sep=';', index=False)

df_char = df_characters.copy()
# we remove the new characters from the df which contains them all
for i in range(len(df_char.Character)):
    if df_char.Character[i] in new_character_3:
        df_char = df_char.drop([i])

df_char = df_char.reset_index(drop=True)

# df_char.to_csv('../Data/temporary_data/Characters_preprocessed.csv', sep=';', index=False)
