==> Notebook/4.additional_features.ipynb <==
from functions import *

data_char = pd.read_csv('../Data/temporary_data/Characters_preprocessed.csv', sep=';', encoding='unicode_escape')
data_char_3 = pd.read_csv('../Data/temporary_data/New_characters.csv', sep=';', encoding='unicode_escape')

# Loyalty

Extraction of all the loyalties:

df_loyalty = data_char[['Character', 'Loyalty']].copy()
df_loyalty.Loyalty = df_loyalty.Loyalty.fillna('')
df_loyalty.Loyalty = df_loyalty.Loyalty.apply(lambda t : t.lower().split(', '))

loyalty_col = []
for i in range(len(df_loyalty)):
    loyalty_list = df_loyalty.Loyalty[i]
    for j in range(len(loyalty_list)):
        loyalty = loyalty_list[j]
        if loyalty not in loyalty_col:
            loyalty_col.append(loyalty)
loyalty_col.remove('')

df_loyalty_3 = data_char_3[['Character', 'Loyalty']].copy()
df_loyalty_3.Loyalty = df_loyalty_3.Loyalty.fillna('')
df_loyalty_3.Loyalty = df_loyalty_3.Loyalty.apply(lambda t : t.lower().split(', '))

loyalty_col_3 = []
for i in range(len(df_loyalty_3)):
    loyalty_list_3 = df_loyalty_3.Loyalty[i]
    for j in range(len(loyalty_list_3)):
        loyalty_3 = loyalty_list_3[j]
        if loyalty_3 not in loyalty_col_3:
            loyalty_col_3.append(loyalty_3)
loyalty_col_3.remove('')

loyalties = list(set(loyalty_col + loyalty_col_3))
loyalties

## HP1 and HP2

loyalty_mat = pd.DataFrame(0, columns=loyalties, index=df_loyalty.Character).rename_axis('Character').reset_index()

for i in range(len(df_loyalty)):
    loyalty_list = df_loyalty.Loyalty[i]
    for j  in range(len(loyalties)):
        loyalty = loyalties[j]
        if loyalty in loyalty_list:
            loyalty_mat[f'{loyalty}'][i] = 1

## HP3

loyalty_mat_3 = pd.DataFrame(0, columns=loyalties, index=df_loyalty_3.Character).rename_axis('Character').reset_index()

for i in range(len(df_loyalty_3)):
    loyalty_list_3 = df_loyalty_3.Loyalty[i]
    for j  in range(len(loyalties)):
        loyalty_3 = loyalties[j]
        if loyalty_3 in loyalty_list_3:
            loyalty_mat_3[f'{loyalty_3}'][i] = 1

# Skills

Extraction of all the skills:

df_skills = data_char[['Character', 'Skills']].copy()
df_skills.Skills = df_skills.Skills.fillna('')
df_skills.Skills = df_skills.Skills.apply(lambda t : t.lower().split(', '))

skills_col = []
for i in range(len(df_skills)):
    skill_list = df_skills.Skills[i]
    for j in range(len(skill_list)):
        skill = skill_list[j]
        if skill not in skills_col:
            skills_col.append(skill)
skills_col.remove('')

df_skills_3 = data_char_3[['Character', 'Skills']].copy()
df_skills_3.Skills = df_skills_3.Skills.fillna('')
df_skills_3.Skills = df_skills_3.Skills.apply(lambda t : t.lower().split(', '))

skills_col_3 = []
for i in range(len(df_skills_3)):
    skill_list_3 = df_skills_3.Skills[i]
    for j in range(len(skill_list_3)):
        skill_3 = skill_list_3[j]
        if skill_3 not in skills_col_3:
            skills_col_3.append(skill_3)

skills = list(set(skills_col + skills_col_3))

## HP1 and HP2

skill_mat = pd.DataFrame(0, columns=skills, index=df_skills.Character).rename_axis('Character').reset_index()

for i in range(len(df_skills)):
    skill_list = df_skills.Skills[i]
    for j  in range(len(skills)):
        skill = skills[j]
        if skill in skill_list:
            skill_mat[f'{skill}'][i] = 1

## HP3

skill_mat_3 = pd.DataFrame(0, columns=skills, index=df_skills_3.Character).rename_axis('Character').reset_index()

for i in range(len(df_skills_3)):
    skill_list_3 = df_skills_3.Skills[i]
    for j  in range(len(skills)):
        skill_3 = skills[j]
        if skill_3 in skill_list_3:
            skill_mat_3[f'{skill_3}'][i] = 1

# New matrices

## HP1 and HP2

my_matrix = pd.read_csv('../Data/Script_matrix.csv', sep=';', encoding='unicode_escape')

my_matrix = my_matrix.merge(loyalty_mat, on='Character')
my_matrix = my_matrix.merge(skill_mat, on='Character')

my_matrix.head()

# my_matrix.to_csv('../Data/Final_matrix.csv', sep=';', index=False)

## HP3

my_matrix_3 = pd.read_csv('../Data/Script_matrix_3.csv', sep=';', encoding='unicode_escape')

my_matrix_3 = my_matrix_3.merge(loyalty_mat_3, on='Character')
my_matrix_3 = my_matrix_3.merge(skill_mat_3, on='Character')

my_matrix_3.head()

# my_matrix_3.to_csv('../Data/Final_matrix_3.csv', sep=';', index=False)
